+++ types/word.ts
@@ -2,7 +2,6 @@ export interface Word {
   id: number;
   word: string;
   phonetic: string;
-  phonics: string;
   chinese: string;
   pos: string;
   example: string;